Move `src/plugins/util/` to `src/util/`.
As the utility methods are not just useful to plugins, move them out of
the `src/plugins/` directory.
