+++ tests/unit/all.ts
@@ -2,7 +2,7 @@ import './main';
 import './plugins/CoreLoadPlugin';
 import './plugins/I18nPlugin';
 import './plugins/InjectModulesPlugin';
-import './plugins/util/i18n';
-import './plugins/util/main';
+import './util/i18n';
+import './util/main';
 import './loaders/css-module-decorator-loader';
 import './loaders/css-module-dts-loader';
