+++ tests/unit/plugins/CoreLoadPlugin.ts
@@ -2,7 +2,7 @@ import { describe, it } from 'intern!bdd';
 import * as assert from 'intern/chai!assert';
 import * as path from 'path';
 import LoadPlugin from '../../../src/plugins/CoreLoadPlugin';
-import { hasExtension, resolveMid } from '../../../src/plugins/util/main';
+import { hasExtension, resolveMid } from '../../../src/util/main';
 import ConcatSource = require('webpack-sources/lib/ConcatSource');
 import NormalModuleReplacementPlugin = require('webpack/lib/NormalModuleReplacementPlugin');
 import Compilation = require('../../support/webpack/Compilation');
