+++ tests/unit/util/main.ts
@@ -1,6 +1,6 @@
 import { describe, it } from 'intern!bdd';
 import * as assert from 'intern/chai!assert';
-import { getBasePath, hasExtension, isRelative, mergeUnique } from '../../../../src/plugins/util/main';
+import { getBasePath, hasExtension, isRelative, mergeUnique } from '../../../src/util/main';
 
 describe('plugins/util/main', () => {
 	describe('getBasePath', () => {
